requirements.txt
slackclient
googletrans
pyjokes